chore: Rename integer variable to fixedPointInt in Fixed class
This commit renames the `integer` variable to `fixedPointInt` in the Fixed class. This change improves the clarity and readability of the code by using a more descriptive variable name.

The purpose of this change is to make the code more self-explanatory and easier to understand for developers working with the Fixed class.