feat: Add Point class to cpp02/ex03
The code changes add the Point class to cpp02/ex03, including the necessary header and source files. This class represents a point in a two-dimensional space and provides functions to get and set the coordinates of the point.

These changes enhance the functionality of the cpp02/ex03 program by introducing the Point class, allowing for more complex calculations and operations involving points as well as Triangle Class