refactor: Add setRawBits function to main.cpp
The code changes in main.cpp add a setRawBits function to the existing code. This function sets the raw bits of an object and logs the new value. This refactor enhances the functionality of the code by introducing a new method for manipulating object data.

Add setRawBits function to main.cpp for manipulating object data