refactor: Add setIdeas method to Brain class
The code changes in Brain.hpp and Brain.cpp add a setIdeas method to the Brain class. This method allows setting new ideas for the Brain object. This refactor enhances the functionality of the code by providing a way to update the ideas stored in the Brain.

Add setIdeas method to Brain class