refactor: Add MateriaSource class and update Makefile
The code changes add the MateriaSource class and update the Makefile in the cpp04/ex03 directory. The MateriaSource class is a new class that implements the IMateriaSource interface. The Makefile has been modified to include the MateriaSource.cpp file in the list of source files. This refactor enhances the codebase by adding a new class and updating the build process.

Add MateriaSource class and update Makefile